### 2025-09-14 #8982 @editorialbot
[PRE REVIEW]: DART: A GUI Pipeline for Aligning Histological Brain Sections to 3D Atlases and Automating Laser Microdissection
**Submitting author:** @mmganant (<a href="http://orcid.org/0000-0002-6600-3410">Manjari Anant</a>)
**Repository:** https://github.com/rk324/Dissecting-Atlas-Registered-Tissue
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
