### 2025-09-18 #8982 @adamltyson
Hi @mmganant, I will take over editorial duties now, and the first step is to find some reviewers. Would you be able to recommend appropriate reviewers for this submission? You can use the [list of potential reviewers](https://reviewers.joss.theoj.org/reviewers) to help. Please provide a list, but don't tag them (with @) so they aren't notified unnecessarily.

Also, although very nicely written, the paper is a bit too long. Our [submission guidelines](https://joss.readthedocs.io/en/latest/paper.html) state that papers should be between 250-1000 words. Would you mind cutting it down? Some of the content could perhaps be added to the documentation instead?

Lastly, one of the references has an invalid DOI. Would you mind taking a look?
