### 2025-09-19 #8982 @rk324
Hi @adamltyson, we've shortened the paper and fixed the invalid DOI. We'll get back to you soon with a list of recommended reviewers.
