### 2025-09-30 #8982 @adamltyson
If everyone is happy and willing, I don't think a submission can be reviewed too well!
