### 2025-09-30 #9109 @editorialbot
[REVIEW]: DART: A GUI Pipeline for Aligning Histological Brain Sections to 3D Atlases and Automating Laser Microdissection
**Submitting author:** @mmganant (<a href="http://orcid.org/0000-0002-6600-3410">Manjari Anant</a>)
**Repository:** https://github.com/rk324/Dissecting-Atlas-Registered-Tissue
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0

@ianhi & @Yannan-Chen & @NicoKiaru & @Julie-Fabre, your review will be checklist based. Each of you will have a separate checklist that you should update when carrying out your review.
