### 2025-10-01 rk324/Dissecting-Atlas-Registered-Tissue#1 @Yannan-Chen
Brain region tree view cut off
Hi,

I noticed that the brain region ROI tree view is partially cut off, while changing the window size doesn't help much. This will make it difficult to see most of the higher-level region names. Could you please take a look at this? Thanks.

This issue is part of the JOSS review https://github.com/openjournals/joss-reviews/issues/9109
