### 2025-10-01 #8982 @carlocastoldi
> 👋 Would you be able to review this submission to JOSS?
> [..] if you're keen to review, I'm sure there will be many more submissions!

Hi @adamltyson, thanks for thinking of me! I'd love to review  in the future and appreciate the offer. However, I'm expecting to be quite busy over the next couple of months and I'm concerned I won't be able to devote the time needed to provide a thorough review.
Please keep me posted on future opportunities, and I'll be happy to consider reviewing then.

On this review, specifically, I think @NicoKiaru point of view is going to be pretty similar to mine. I have my full trust in him!
