### 2025-10-14 rk324/LDM-brain-region-excision#2 @Julie-Fabre
Installation and running steps for non-Windows users
Hi,
I found your installation instructions on the main readMe, but could you point to where to describe how to compile/launch the app when *not* using the 'Windows  pre-compiled' method?
Thank you very much!
This issue is related to the JOSS review: https://github.com/openjournals/joss-reviews/issues/9109
