### 2025-10-14 rk324/LDM-brain-region-excision#3 @Julie-Fabre
Automated tests
Are there any automated tests for this software (and if so, could you point me towards them)? Thanks a lot!
This issue is related to the JOSS review: https://github.com/openjournals/joss-reviews/issues/9109
