### 2025-10-20 rk324/Dissecting-Atlas-Registered-Tissue#3 @rk324
@Julie-Fabre The tests are located at [dart/test](https://github.com/rk324/Dissecting-Atlas-Registered-Tissue/tree/main/dart/test). You can run them with the command `pytest --disable-warnings` or `python -m pytest --disable-warnings`.  I've updated the README to show these commands. Thanks for the feedback!
