### 2025-10-20 rk324/Dissecting-Atlas-Registered-Tissue#1 @rk324
@Yannan-Chen Thanks for pointing this out! We've modified the window to have an adjustable width and a horizontal scrollbar.
