### 2025-10-21 rk324/Dissecting-Atlas-Registered-Tissue#2 @Julie-Fabre
Hi @rk324 thanks a lot! Could you add a more detail - namely once users have cloned and setup the repo/atlases - how do you call DART?
