### 2025-10-22 rk324/Dissecting-Atlas-Registered-Tissue#4 @NicoKiaru
Mouse Slice Data - where does the data come from ?
Hello,

I see that you provide an example image, which is great to test the software. However could you mention where the data is coming from and if specify the animal license used if applicable ? Is it an original dataset ? This is for the animal research policy of joss: https://joss.readthedocs.io/en/latest/policies.html#joss-policies

A small text file in the image sample directory would work.

issue linked to joss review: https://github.com/openjournals/joss-reviews/issues/9109
