### 2025-10-22 rk324/Dissecting-Atlas-Registered-Tissue#5 @NicoKiaru
Clarify contributing guidelines ?
The current contributing guidance states:

> "To contribute to the software, please email manjari.anant@gmail.com and rishikoneru2005@gmail.com."

- Are pull requests accepted directly, or must contributors contact you first?
- What is the preferred contribution workflow?
- Are there coding standards or guidelines contributors should follow?

**Suggestion:**
Consider adding a CONTRIBUTING.md file that clarifies whether PRs are welcome and under what circumstances

This will make it easier for potential contributors to engage with the project

linked to review https://github.com/openjournals/joss-reviews/issues/9109
