### 2025-10-22 rk324/Dissecting-Atlas-Registered-Tissue#6 @NicoKiaru
Improve data availability: Replace Google Drive with archival repository
Currently, atlas data is distributed via Google Drive links. While convenient, this raises concerns about long-term sustainability:

1. **Link longevity:** Google Drive links can break if accounts change, files are moved, or sharing settings are modified
2. **Reproducibility:** research best practices favor permanent, versioned data repositories
3. **Discoverability:** Data in institutional repositories is indexed and citable

**Recommendations:**

**Option 1 - Zenodo:**
- Upload atlas data to [Zenodo](https://zenodo.org/)
- Provides DOI for citation
- Permanent, versioned storage
- Free for academic use
- Widely used in research software

**Option 2 - Leverage existing repository:**
Consider depending on [BrainGlobe Atlas API](https://brainglobe.info/documentation/brainglobe-atlasapi/index.html), which:
- Already handles Allen Brain Atlas data
- Manages versioning and downloads
- Reduces maintenance burden
- Improves interoperability with other tools

linked to review https://github.com/openjournals/joss-reviews/issues/9109
