### 2025-11-18 #9109 @Julie-Fabre
Hi all,
I am done with my review. Good job! Overall, I think this software satisfies JOSS requirements: it is novel and corresponds to a current need in the field. I have been able to test the installation and aligning but I don’t have access to an LMD microscope so could not test that part.

@NicoKairu has already done an excellent review job and I agree with many of his points, namely:
Ideally this package would be available on PyPi
It would be ideal to have the atlases available elsewhere than google drive - for instance making use of the brainglobe repository as highlighted in issue#6

Finally, I agree with @[NicoKiaru](https://github.com/NicoKiaru) regarding performance considerations (especially issue #7). These are out of the scope of this current review in my opinion but should be considered to make this package more robust, fast and user-friendly.

Good job!
Julie
