### 2025-12-07 rk324/Dissecting-Atlas-Registered-Tissue#6 @mmganant
Hello, thanks for this suggestion! We have uploaded the atlas data to Zenodo (https://zenodo.org/records/17849564), and changed the Google drive link in our README to Zenodo.
