### 2025-12-07 #9109 @mmganant
We thank the reviewers for testing the software and for their constructive suggestions. We have made the following updates:

**1) Contributing guidelines**

- Added a CONTRIBUTING file covering issue/PR workflow, coding style, and test instructions.
- Path: [https://github.com/rk324/Dissecting-Atlas-Registered-Tissue/blob/main/CONTRIBUTING.md](https://github.com/rk324/Dissecting-Atlas-Registered-Tissue/blob/main/CONTRIBUTING.md)

**2) Public data deposit**

- Uploaded the dataset to Zenodo; the README and manuscript now cite the release.
- Zenodo: [https://zenodo.org/records/17849564](https://zenodo.org/records/17849564)

**3) Animal research compliance**

- Explicitly reported institutional approvals (JHU ACUC protocols MO20M376 and MO23M346).
- Provided ARRIVE 2.0 (Essential 10) details.
- Path: [https://github.com/rk324/Dissecting-Atlas-Registered-Tissue/blob/main/demo_images/ARRIVE_Essential10.txt](https://github.com/rk324/Dissecting-Atlas-Registered-Tissue/blob/main/demo_images/ARRIVE_Essential10.txt)

**Packaging on PyPI**

We appreciate the suggestion to publish on PyPI. At present, the software is distributed via a Windows executable and source on GitHub, which we believe complies with JOSS packaging guidelines. Given the dependency on VisuAlign and a large atlas directory, a lightweight PyPI package would require additional restructuring, which we can integrate into a future release.

**Scope of integrations**

We agree that DeepSlice compatibility and Napari visualization would be useful. To keep this submission focused on core functionality, we believe these are out of scope for the initial JOSS paper. We plan to explore an optional DeepSlice component and a minimal Napari plugin in subsequent releases.
