### 2026-01-12 #9109 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
